+++ packages/shared/components/ColorPicker.svelte
@@ -1,7 +1,7 @@
 <script lang="typescript">
     import { Text, Icon, Tooltip } from 'shared/components'
     import { isBright } from 'shared/lib/helpers'
-    import { clickOutside } from 'shared/lib/actions'
+    import { clickOutside } from '@core/utils'
     import { localize } from '@core/i18n'
     import { AccountColors } from '@core/account'
