+++ packages/shared/lib/core/utils/ui.ts
@@ -1,3 +1,6 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
+/* eslint-disable @typescript-eslint/explicit-module-boundary-types */
+
 import { Event } from '@lib/typings/events'
 
 import { Element } from './interfaces'
@@ -25,5 +28,36 @@ export function debounce(callback: () => void, wait = 500): (...args: unknown[])
         const context = this
         clearTimeout(_timeout)
         _timeout = setTimeout(() => callback.apply(context, args), wait)
+    }
+}
+
+/**
+ * Dispatch event on click outside of node
+ * source: https://svelte.dev/repl/0ace7a508bd843b798ae599940a91783?version=3.16.7
+ */
+export function clickOutside(node: any, options?: { includeScroll }): { destroy } {
+    const handleClick: (event: MouseEvent) => void = (event) => {
+        if (node && !node.contains(event.target) && !event.defaultPrevented) {
+            node.dispatchEvent(new CustomEvent('clickOutside', node))
+        }
+    }
+
+    const handleScroll: () => void = () => {
+        node.dispatchEvent(new CustomEvent('clickOutside', node))
+    }
+
+    document.addEventListener('mousedown', handleClick, true)
+
+    if (options?.includeScroll) {
+        document.addEventListener('scroll', handleScroll, true)
+    }
+
+    return {
+        destroy(): void {
+            document.removeEventListener('click', handleClick, true)
+            if (options?.includeScroll) {
+                document.removeEventListener('scroll', handleScroll, true)
+            }
+        },
     }
 }
